#12, #37 added service class and tests for CoopDetails

+++ Cooperator-Backend/src/main/java/ca/mcgill/cooperator/service/CoopDetailsService.java
@@ -22,14 +22,17 @@ public class CoopDetailsService {
 	@Transactional
 	public CoopDetails createCoopDetails(int payPerHour, int hoursPerWeek, EmployerContact ec, Coop c) {
 		StringBuilder error = new StringBuilder();
-        if (payPerHour == 0) {
-            error.append("Pay Per Hour cannot be empty! ");
+        if (payPerHour < 0) {
+            error.append("Pay Per Hour is invalid! ");
         }
-        if (hoursPerWeek == 0) {
-            error.append("Hours Per Week cannot be empty! ");
+        if (hoursPerWeek <= 0) {
+            error.append("Hours Per Week is invalid! ");
         }
         if (ec == null) {
-        	error.append("Employer Contact cannot be null!");
+        	error.append("Employer Contact cannot be null! ");
+        }
+        if (c == null) {
+        	error.append("Coop cannot be null!");
         }
         if (error.length() > 0) {
             throw new IllegalArgumentException(error.toString().trim());
@@ -57,16 +60,85 @@ public class CoopDetailsService {
 	}
 	
 	@Transactional
+	public CoopDetails getCoopDetails(int id) {
+		CoopDetails cd = coopDetailsRepository.findById(id).orElse(null);
+        if (cd == null) {
+            throw new IllegalArgumentException("Coop Details with ID " + id + " does not exist!");
+        }
+
+        return cd;
+	}
+	
+	@Transactional
 	public List<CoopDetails> getAllCoopDetails() {
 		return ServiceUtils.toList(coopDetailsRepository.findAll());
 	}
 	
 	@Transactional
 	public CoopDetails deleteCoopDetails(CoopDetails cd) {
+		if (cd == null) {
+			throw new IllegalArgumentException("Coop Details to delete cannot be null!");
+		}
 		Coop c = cd.getCoop();
 		c.setCoopDetails(null);
 		coopRepository.save(c);
+		
+		EmployerContact ec = cd.getEmployerContact();
+		List<CoopDetails> details = ec.getCoopDetails();
+		details.remove(cd);
+		ec.setCoopDetails(details);
+		employerContactRepository.save(ec);
+		
 		coopDetailsRepository.delete(cd);
 		return cd;
 	}
+	
+	@Transactional
+	public CoopDetails updateCoopDetails(CoopDetails cd, int payPerHour, int hoursPerWeek, EmployerContact ec, Coop c) {
+		StringBuilder error = new StringBuilder();
+        if (payPerHour < 0) {
+            error.append("Pay Per Hour is invalid! ");
+        }
+        if (hoursPerWeek <= 0) {
+            error.append("Hours Per Week is invalid! ");
+        }
+        if (ec == null) {
+        	error.append("Employer Contact cannot be null! ");
+        }
+        if (c == null) {
+        	error.append("Coop cannot be null!");
+        }
+        if (error.length() > 0) {
+            throw new IllegalArgumentException(error.toString().trim());
+        }
+        
+        cd.setPayPerHour(payPerHour);
+        cd.setHoursPerWeek(hoursPerWeek);
+        cd.setEmployerContact(ec);
+        cd.setCoop(c);
+        
+        coopDetailsRepository.save(cd);
+        
+        List <CoopDetails> employerCoopDetails = ec.getCoopDetails();
+        boolean employerContains = false;
+        for (CoopDetails coopDetails : employerCoopDetails) {
+        	if (coopDetails.getId() == cd.getId()) {
+        		int index = employerCoopDetails.indexOf(coopDetails);
+        		employerCoopDetails.set(index, cd);
+        		employerContains = true;
+        	}
+        }
+        if (employerContains == false) {
+        	employerCoopDetails.add(cd);
+        }
+        ec.setCoopDetails(employerCoopDetails);
+        
+        c.setCoopDetails(cd);
+        
+        employerContactRepository.save(ec);
+        coopRepository.save(c);
+        
+        return coopDetailsRepository.save(cd);
+	}
+	
 }
